Add Debug config to embed assemblies into APK for Android
A new <PropertyGroup> has been added to the FinanceTracker.Mobile.csproj file. This property group is conditioned to apply when the configuration is 'Debug', the target framework is 'net8.0-android', and the platform is 'AnyCPU'. Within this property group, the <EmbedAssembliesIntoApk> property is set to True, embedding assemblies into the APK during the build process for this specific configuration. This simplifies the deployment process for debugging.